requirements.txt
alembic==1.12.1
anyio==4.0.0
async-timeout==4.0.3
bidict==0.22.1
build==1.1.1
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
deprecation==2.1.0
distro==1.8.0
docopt==0.6.2
exceptiongroup==1.1.3
fastapi==0.96.1
gotrue==2.4.1
greenlet==3.0.1
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
Jinja2==3.1.2
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
packaging==23.2
pipdeptree==2.16.1
pipreqs==0.4.13
platformdirs==3.11.0
postgrest==0.11.0
psutil==5.9.6
pydantic==1.10.13
Pygments==2.16.1
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.0
python-multipart==0.0.5
python-socketio==5.10.0
realtime==1.0.0
redis==4.6.0
reflex==0.4.5
reflex-hosting-cli==0.1.9
requests==2.31.0
rich==13.6.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.28
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.14
starlette==0.27.0
starlette-admin==0.9.0
storage3==0.6.1
StrEnum==0.4.15
supabase==2.2.1
supafunc==0.3.3
tabulate==0.9.0
tomli==2.0.1
typer==0.4.2
typing_extensions==4.8.0
urllib3==2.2.1
uvicorn==0.20.0
watchdog==2.3.1
watchfiles==0.19.0
websockets==10.4
wrapt==1.15.0
wsproto==1.2.0
yarg==0.1.9
